Consolidate Brokers
This commit consolidates all of the brokers, while making it easier to
add further events, and correlated events. It also switches from a
purely event driven system to a window-with-configurable-number-of-errs
and a counter. Added a dependency on pytest for running this:

pytest [-s] src/world_broker.py

The -s flag will print otherwise suppressed output(for printf debugging)
.