Fix leader election with ReceiveDrop
There were two problems causing leader election to fail during ReceiveDrop
events. The first was that followers did not keep track of who the voted for,
and thus could vote for more than one candidate. The second was that followers
would try to directly transition into leaders when they got a
RequestVoteResponse - thus previous candidates who had heard from a leader or a
higher term candidate and transitioned to a follower when a RequestVoteResponse
was still in the process of being sent were able to convert into a leader once
they received enough (stale) votes.

This commit also adds more logging information to the broker log. I'm not 100%
sure how this is meant to be used, so please correct me if I should be using it
differently.